requirements.txt
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
scikit-learn>=1.3.0
xgboost>=1.7.0
lightgbm>=3.3.0
catboost>=1.2.0
imbalanced-learn>=0.11.0
statsmodels>=0.14.0
prophet>=1.1.0
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
PyYAML>=6.0
joblib>=1.3.0
tqdm>=4.65.0
pytest>=7.4.0
